e031: improve string concatenation example.
Instead of allocating a `String` and dealing with the UTF-8 checking it
entails, use `std::slice::concat` to merge the two slices of `u8` we
already have. This gives us an owned `CString` we can then convert back
into a `c_char` and send across the FFI boundary. If we were going to use
it inside Rust, accepting the cost of the conversion would make sense,
but since we are only sending it right back across the boundary, there's
no need to pay that cost.